test: 收編孤兒打包遷移檢查為可收集測試，缺打包證據時明確 skip／收编孤儿打包迁移检查为可收集测试，缺打包证据时明确 skip／adopt the orphan packaged-migration checker as a collected test with an explicit evidence-gated skip／孤児のパッケージ移行チェックを収集対象テストに編入し、証拠欠如時は明示的に skip
